don't use real qrencode in tests
There are subtile differences in the codes generated # by different
versions of qrencode that break our tests (my local system vs. Travis
CI for example).

Because of all the error corrections in a QR code, slightly different
results are valid, but using diff(1) to check the result fails because
of the differences.

Replace qrencode by a simple shell function that transforms the input
in a non-ambigious way.